fix: refactor GITHUB_REF handling to use temp ref
The --porcelain flag is very nice, but unfortunately, it's not present
on older git versions. Since the client git version isn't controlled or
managed by us, we should try and play nice with it, and not use any
features that are too new.

In this new approach, we create a new "refs/dagger/..." ref, and then
update that (to avoid making changes to any other refs in the repo), and
then delete it at the end.

Signed-off-by: Justin Chadwell <me@jedevc.com>