engine: make client IDs fully random again (#7335)
* engine: make client IDs fully random again

A few releases ago, client IDs were changed to be digests of the
dagql.CurrentID in order to remove the number fields in our ftp_proxy
hack and to consolidate client ID w/ module caller digest.

This is not entirely working as expected though (detailed below).
Fortunately the new custom worker gives us a non-hacky path to set
whatever IDs we want in containers without worrying about cache busting
or the various problems caused by the ftp_proxy hack, which has been
gone entirely as of last release.

This updates the engine to make client IDs random strings again.

The problem we're hitting is that when there are inputs to a Function of
type Secret, the Functions seem to always run without any caching no
matter what, even in the same dagger server. And even though before this
client IDs would be the same, each client still generated it's own
random secret token and thus conflicted with each other. Making client
IDs random again fixes it because the secret tokens no longer mismatch.

However, the fact that we are not getting cache hits when invoking the
same function repeatedly when there is an input of type Secret is still
a mystery at this time. I confirmed that we do get the expected caching
for other input types (string, Container, etc.), it's somehow only
happening for Secret inputs.

I'm going to open a separate issue with more details for figuring out
what's happening there.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

* scope function call execop cache key to dagql ID

otherwise we actually cache **too much** and the call inputs/parent val
etc. don't count for the cache key and every function in a module
returns the same value.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

* chore: add changelog

Signed-off-by: Justin Chadwell <me@jedevc.com>

---------

Signed-off-by: Erik Sipsma <erik@sipsma.dev>
Signed-off-by: Justin Chadwell <me@jedevc.com>
Co-authored-by: Justin Chadwell <me@jedevc.com>