chore: remove noVCSSuffix
Remove noVCSSuffix to allow the .git suffix on the SCM matched via a regex.
The underlying reason is to accept inputs with and without this VCSSuffix.

This is useful, as in order to resolve the root of a GitLab private repo, we need to specify the root of the repository with `.git` for the function to piggyback on the ping command (in our case: `git ls-remote -q`)

Signed-off-by: grouville <guillaume@dagger.io>