core: plumb server ID via executionMetadata
we only sometimes need the server ID as an env var in order to
invalidate cache, so it's a lot cleaner+nicer to just pass it through
the exec metadata than rely solely on an env var to pass it around.

Also makes it easier for other code (i.e. service + dockerfile frontend)
to set it without jumping through the hoops of stringly-typed env vars.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>