core: fix custom CA certs in modules + add integ test coverage
A recent commit that changed client IDs to be random again accidentally
made it so that commands run to update CA certificates (when a custom
engine has those installed) hit errors about client tokens not matching
when the exec being run was nested, which includes module function
calls.

This was due to the fact that those commands run in the same container
with the same client ID but the secret token was generated by the client
each time it ran, which resulted in mismatches.

The fix here just updates those extra CA commands to never be nested
execs since there's no purpose in that anyways and it's nothing but
overhead.
* We just unset the relevant env var for those commands, which is a bit
  ugly but not any more than the existing ugliness. Fortunately the
  other ongoing effort to clean all this up should help a lot here.

This was missed because I forgot to include integ tests that invoke
module functions amongst all the other many tests that were added for
custom CA support. Those have been added now for go, python and
typescript.

When adding the test for typescript functions, I also learned that
nodejs will by default only use a CA bundle compiled into the node
binary. So this also updates the typescript SDK to tell nodejs to use
the system CAs instead (which also making sure they include the default
Mozilla CA set, which is what nodejs is compiled with anyways).

Signed-off-by: Erik Sipsma <erik@sipsma.dev>