core: remove ONLCR dance from shim
I was unable to replicate any problems with linebreaks not displaying
correctly once this was removed from the shim, even though I verified
that stty showed `-onlcr` in `dagger call ... terminal` containers.

I *suspect* that the more recent move in the CLI to just use the raw
terminal rather than an emulator results in the ONLCR of *my* terminal
(as a client) mattering rather than that of the container.

Worst case and this is needed after all, we can use OCI hooks for it
instead of the shim.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>