core: migrate networking and OTEL related code to executor
The healthchecks and tunneling that used to require a full-blown
container that short-circuited to special shim commands can now just be
done in the engine directly.

That is done here by adding a util for executing go funcs inside the
network namespace of existing containers (or creating a new raw network
namespace in the case of the tunelling, since we just need an ip and
listener there.

The util hides the tricky details of needing to lock goroutines so
callers can just run the func they want.

It works by maintaining a pool of goroutines that will Setns into the
container's network namespace when a job comes in over a channel. It
also will stay locked in the namespace up to a timeout (currently just a
second).
* The reasoning here is that we don't want to consume an entire OS
  thread for too long but also want to avoid rapidly entering and
  leaving the namespace for e.g. health checks.

The OTEL related code is grouped in here since the bulk of the work is
setting up the TCP proxying of the otel unix sock, which relies on the
above mechanisms.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>