fix: hide remotes.docker.resolver spans (#7386)
Before this, we would display all the containerd/buildkit spans that
relate to pulling an image. These are *super* unhelpful by default, so
we want to hide them in the TUI (unless the parent command errors).

We already have this concept today with `telemetry.Encapsulate`.
However, this is set on the parent (in our case, `Container.from`),
which isn't quite what we want. Firstly, this would silence all output
from this function (even if later we would want to add some more useful
output of our own). Secondly, in some builds, these requests might not
necessarily even occur as part of `Container.From` - we might end up
lazily pulling the image later.

So, we add a new `telemetry.Encapsulated` property - this is exactly the
same as the above, but is instead set on a child span. We can set this
using our new buildkit tracing interceptor, and only set it on the
buildkit spans that we care about.

Signed-off-by: Justin Chadwell <me@jedevc.com>