engine: Remove oci-max-parallelism
This is not an option that we should support. Justin puts it best:

    I think with (Dagger) Modules, this effectively means that
    `max-parallelism` is pretty much just completely broken - it just fits
    very badly with the idea of nested containers.

    ... but to me this kind of highlights that we need a better way of
    constraining resources - max-parallelism is just really not fit for
    the shape that pipelines tend to take these days. Maybe we should
    provide some cgroup config options instead (if the end desire is to
    constrain CPU usage/etc).

While this can still be configured if a custom Engine toml config is
provided, it means that users went out of their way to have this. It's
hard on purpose and we don't recommend using this setting anymore.

This also removes it from the Helm chart.

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>