core: build dumb-init and use as init for containers
The removal of the shim was actually a breaking change due to subtle but
important differences in how Linux treats pid 1 processes.

Namely, pid 1 will not get any default signal handlers and instead must
explicitly register them. For example, without the shim our integ tests
that ran python -m http.server stopped responding to SIGTERM.

Not having an init also made it easier for zombie processes to lurk
while the container was still running.

To avoid this, we now build dumb-init and package it into the engine
container.

It's extremely lightweight (20kb statically linked and extremely minimal
in what it does) and is also possible make a configuration option on the
Container API for users who don't want it (though it should remain the
default). This is as opposed to our shim which was much heavier and also
was not optional until its functionality migrated to the executor.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>