executor: fix case of looking up unqualified service name
The executor does not Setns into the mount namespace of containers, only
the network namespace, so the Go resolver doesn't have access to the
/etc/resolv.conf with search domains specific to the container, which
resulted in service hostnames failing to be resolved.

While adding support for mount namespace entering is feasible, it would
require we do a dance to pre-create a mount namespace and set it in the
spec. It's a lot simpler to just check the search domain directly in our
code. If any future use cases call for entering mount namespaces we can
revisit and just use that instead.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>