ci: Add extra flags for Engine & CLI tests
`--parallel` now defaults to the number of CPUs. Before, this was
hard-coded to `16` which was set the number of CPUs in the larger GitHub
Runners, the `ubuntu-22.04-16c-64g-600gb` instance size specifically. If
this new default value does not work well on the larger CI runners (they
have 48CPUs & we run multiple pipelines on them at the same time), we
can specify this value explicitly in the workflows.

`--timeout` defaults to `30m`. This was changed not that long ago - the
value was set to 1h when using race. The bottom line is that none of our
test suites should take more than 10 minutes. This is something that we
are actively working towards. For now, until the pipeline becomes more
reliable, we will stick with the current 30m default. When running
locally with fewer CPUs, or when hunting deadlocks, experience
recommends a value of 180m.

`--failfast` is a new one. It sets `--max-fails=1` in `gotestsum`. FTR:
https://github.com/gotestyourself/gotestsum/blob/e9677fb405a5c0dc0b9d7217a7c73c3a20c5cb66/cmd/testdata/gotestsum-help-text#L20

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>