fix(cli): don't validate flags when requesting `--help`
The `--help` flag is now global (interspersed). So the help will no longer be shown next to the command where it’s defined if there’s more to the right. It’s position in the chain no longer matters, it will be the same as adding it to the end. This was needed to look ahead and skip some validations while that flag wasn’t parsed yet, and while it was possible to preserve the previous behavior, making it global seems more consistent with CLI conventions, besides being simpler.

There's also some refactoring to make the logic easier to follow and encapsulate some responsibilities better in functions.

Signed-off-by: Helder Correia <174525+helderco@users.noreply.github.com>