cli: don’t show functions that can’t be called
Most common reason for not being able to call a function is unsupported flags, but if there's unsupported return types as well, they should be added too.

This change excludes functions with **any** unsupported flag, however, maybe it’s best to do it only for required flags, and keep the function but skip adding the optional arguments that aren’t supported.

Signed-off-by: Helder Correia <174525+helderco@users.noreply.github.com>