engine: centralize engine server-side setup
There's an enormous amount of buildkit+containerd entities like
snapshotters, various cache dbs, the solver, worker+executor, etc.
which are all important to understand, but previously they were all
setup all over the place which was extremely confusing to follow.

This consolidates them all to be setup and stored in one place. There's
quite a bit there, but at least you don't have to search far and wide to
know what exists and how it's configured.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>