ci: add sdk all group
When we removed mage, we removed the easy helpers that allowed
generating/linting/etc on all SDKs. That meant that making API changes
became much more fiddly.

This patch restores the "all" group, but this time directly in the
dagger module. One day, we might remove this with a script, but we don't
have that right now.

Signed-off-by: Justin Chadwell <me@jedevc.com>