ci: Add extra flags for Engine & CLI tests (#7416)
* ci: Add extra flags for Engine & CLI tests

`--parallel` now defaults to the number of CPUs. Before, this was
hard-coded to `16` which was set the number of CPUs in the larger GitHub
Runners, the `ubuntu-22.04-16c-64g-600gb` instance size specifically. We
still want `16` in CI - we set this explicitly in the GitHub workflows - but are 
likely to want less locally where some of us only have `8` or `10` CPUs.

`--timeout` defaults to `30m`. This was changed not that long ago - the
value was set to 1h when using race. The bottom line is that none of our
test suites should take more than 10 minutes. This is something that we
are actively working towards. For now, until the pipeline becomes more
reliable, we will stick with the current 30m default. When running
locally with fewer CPUs, or when hunting deadlocks, experience
recommends a value of 180m.

`--failfast` is a new one. It sets `--max-fails=1` in `gotestsum`. FTR:
https://github.com/gotestyourself/gotestsum/blob/e9677fb405a5c0dc0b9d7217a7c73c3a20c5cb66/cmd/testdata/gotestsum-help-text#L20

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>

* ci: Default Engine & CLI test parallel to 16 in CI

If it defaults to 64 CPUs, as it did here, it's more likely to fail:
https://dagger.cloud/dagger/traces/8a089c6b4c74959d5a3aef1d8aeded74#b2d066943ad6dbc4:L1051

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>

---------

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>