fix: check error on normalizeCallerLoadedSource
`normalizeCallerLoadedSource` function can segfault if `withContextDirectory` returns an error, as we do not check for the returned error on the `withContextDirectory`.

The following dagql calls can be executed: on the `dagger init`, for example, the `withName` gets called on my tests, and trigger the segfault.

It is currently never triggered, as there are two potential edge cases in which we return an error on `withContextDirectory`:
1. when the modulesourcekind is not local
2. when we fail to load the dir

Signed-off-by: grouville <guillaume@dagger.io>