feat(parseRefString): extend parseRefString local / remote ref logic
Previously, local / git refs were estimated using such heuristic:
- no version specified AND ref starting with "github.com"

However, as we are now extending the logic to other CI providers, this
needs to be updated.

The current logic is extended to stat if the dir is present ; or if the
ref starts with "/", ".", "./".

This shall enable handling subdirectories without local path prefixes
nor absolute path.

However, in order to keep the current logic in place, make sure that the
sourceRootRelPath starts with "./", in the resolveFromCaller func. This enables
 us to keep the rel path correct whithout changing the logic

Signed-off-by: Guillaume de Rouville <guillaume@dagger.io>