ci: build dumb-init for target platform and add verification
dumb-init was accidentally always being built for the host platform,
which meant if it was built on an x86 host the published arm64 image
would incorrectly be x86.

This fixes that and also adds general purpose verification of all the
binaries we build checking that they are of the expected target
platform, to reduce the chances of this happening again.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>