cli: convert Windows paths to / slash format
Before this, file+dir arg values to dagger call on a Windows client were
not converted from `\` format to `/` format, which resulted in calls
like `filepath.Split` in the Linux engine container to not work as
expected.

Now the CLI calls `filepath.ToSlash` before sending paths, which *helps*
handle the paths better in the engine container; specifically enough to
avoid at least one bug that resulted in the engine trying to find a path
at e.g. `/C:\Foo\Bar`.

To be clear, this is probably not the end of our need for improved
Windows path handling, but may be enough to at least squash this one
bug.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>