session: ignore SIGPIPE, don't panic
if either of these happen ther CLI-side span won't be finished and
telemetry won't drain.

The same fix was applied recently for stdout, but I observed SIGPIPE
happening on write(2) - stderr. I suspect this could still happen for
stdout. My guess is it happens when the parent process goes away without
waiting for the process. I couldn't find where that's happening (if it
is happening), so I think the safest bet is to just drop the signal,
since we should be shutting down around then anyway with stdin also
being closed; might as well do it cleanly.

Also, I didn't see the panic(err) become an issue, but it happens in a
goroutine, which AFAIK is un-recoverable and won't run defers. Better to
just log to stderr.

Signed-off-by: Alex Suraci <alex@dagger.io>