fix(ts-sdk) rollback to npm to ignore packageManager field
#7096 introduced a regression by using `yarn install --production`.
If the field `packagaManager` is set, it might lead to failure because
yarn post v1.22.11 doesn't support the `--production` flag anymore,
it has been removed in favor of `yarn workspaces focus --production`.

However, depending on the yarn major version (v2, v3, v4), the options
are different but also the results and the way they handle
subdependencies.

It seems `npm` is the only package manager that actually force install
all the sub dependencies, which ensure the TS SDK runtime properly
works.

Signed-off-by: Tom Chauveau <tom@epitech.eu>