feat: add parseRefString test
Implements the foundation to test parseRefString without side-effects. In
order to do so, we rely on a interface to mock buildkit's stat.

Contrary to the integration tests, this test suite is used to give
insights on how parseRefString behaves on public repos, and how it returns
informations later used by TestSplitRootAndSubdir

Signed-off-by: Guillaume de Rouville <guillaume@dagger.io>