simplify and slightly optimize blobification of schema json
Noticed that we now only need to generate a json file blob for the
version of the schema presented to modules, not both that and the
unaltered one, so I simplified that.

Also realized that it's highly unnecessary to compress these blobs in
the content store, so I added a param to DefToBlob for controlling that
(leaving the other existing usages as Zstd).

I did this because I saw some tests were slow, but on re-run they were
fasted so was probably just noise. But the commit is a simplification
overall so pushing anyways.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>