core: reduce module ID size by not including full schemas (#7549)
* core: reduce module ID size by not including full schemas

Before this, we were passing full introspection schema JSONs as strings
as args to non-Go SDKs, which resulted in them being included in the
full IDs for non-Go modules.

Those strings got huge and could appear multiple times in a single ID,
which inflated their size, and also ended up in progress output (I think
the source of `ETOOBIG`?) + telemetry.

This has been a known issue for a while but not harmful enough to be
prioritized for a fix. However, my other PR that refactors a bunch of
stuff around sessions ended up hitting this problem hard since it
required including Module+FunctionCall IDs as HTTP headers, but then
hitting limits on the max size of HTTP headers (go defaults to 1MB,
which is pretty reasonable+generous).
* Some headers that previously ended up summing to 1.7MB now max out at
  70kb (still too high but I think there's other issues to fix
  separately).

The fix here changes the introspection json to be passed as type File
rather than string. The File is created using the existing
blob-functionality so it's all nice and content hashed.

As I updated the SDK implementations to account for this (technically a
small breaking change), I also noticed that all of them turned the
string into a File anyways internally, so this seems like an all around
better approach for everyone involved.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

* simplify and slightly optimize blobification of schema json

Noticed that we now only need to generate a json file blob for the
version of the schema presented to modules, not both that and the
unaltered one, so I simplified that.

Also realized that it's highly unnecessary to compress these blobs in
the content store, so I added a param to DefToBlob for controlling that
(leaving the other existing usages as Zstd).

I did this because I saw some tests were slow, but on re-run they were
fasted so was probably just noise. But the commit is a simplification
overall so pushing anyways.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

* add changie

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

---------

Signed-off-by: Erik Sipsma <erik@sipsma.dev>