feat: rework cloud progress printing
For pretty progress, we add a new little message panel at the bottom
containing a "Full trace at ..." message.

For plain progress, in addition to the above, we also add a little fake
virtual log to the "connecting to engine" span, so that it's available
from the very start of the trace.

Signed-off-by: Justin Chadwell <me@jedevc.com>