ts: correctly restore env vars after test run
One of the typescript tests was overriding env vars and then deleting
them afterwards, which could result in the original values disappearing.

I can't tell if this was an actual problem in practice at the moment,
but seems safer in the long run to do this in case we ever run these
tests in a nested exec, in which case this would have broken them.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>