ts: pin tsx version installed for modules
tsx pushed a release an hour ago that caused all Typescript modules to
break out of band with errors like:
```
Transforming JavaScript decorators to the configured target environment
("node21.3.0") is not supported yet
```

It broke out-of-band because our install of tsx was not pinned to any
version. I pinned it now to the previous patch release and things work
again.

Hopefully there's some better way of pinning this version since it's
currently just a direct `npm install` call and not in a lock file, but
this at least fixes the problem in the immediate term.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>