Various TS + general test fixes + improvements (#7571)
* ts: use multiplatform index for bun image

The image sha for bun was pointing to an amd64 image, which caused
failures on other platforms. Now it points to the multiplatform manifest
so the right platform gets chosen.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

* tests: fill in missing Parallel calls

Noticed there were a lot of module related integ tests with missing
calls to Parallel where it seemed safe to do so. Given that the actual
concurrency is limited, it seems at worst harmless and at best a
performance improvement to run these in parallel.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

* ts: fix missing awaits in typescript sdk tests

There were a few calls to connect that were not being awaited, which
caused the test to just exit with success instantly while callbacks
still ran async in the background.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

* ts: upgrade bun version to avoid segfaults

While running typescript tests locally I was getting lots of really
crazy errors that caused bun to hit SIGSEGV address-out-of-bounds,
SIGABRTs, etc.

Googling around led to GH issues that said these problems were fixed in
later versions, so I bumped us up to 1.1.12 and have not hit any more of
these.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

* ts: correctly restore env vars after test run

One of the typescript tests was overriding env vars and then deleting
them afterwards, which could result in the original values disappearing.

I can't tell if this was an actual problem in practice at the moment,
but seems safer in the long run to do this in case we ever run these
tests in a nested exec, in which case this would have broken them.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

---------

Signed-off-by: Erik Sipsma <erik@sipsma.dev>