engine/client: don't use immediate shutdown
otherwise dagger up tests don't see the 'start ...' span complete.

this can likely just be removed now; it was added at a time when we had
to be durable to a `kill -9`ed process because its spans would hold the
draining up, but now we drain when the parent span exits.

Signed-off-by: Alex Suraci <alex@dagger.io>