Introduce better versioning compatibility guarantees (#7031)
* hack: add legacy label conversion for old clients

This allows old clients like dagger v0.10.3 to still connect to and use
dagger engine v0.11.0. While this *isn't* a strict requirement, it's
nice to not break this if we don't have to, and based on a couple tests,
it seems like this was the only blocker for that.

This is needed so we can determine legacy client versions.

Signed-off-by: Justin Chadwell <me@jedevc.com>

* fix: ensure client fails out when session call fails

If the main session call that was supposed to register the server fails
out, there can be no recovery! No number of retries will help, so we
should cancel the context.

Additionally, we can manipulate the error handling logic to ensure that
when the context is cancelled for this reason, that we return a
*relevant* error - we can just take the last error, this is good enough.

Signed-off-by: Justin Chadwell <me@jedevc.com>

* feat: introduce minimum versioning requirements

With this change, we finally introduce minimum version requirements, in
two parts:
- Clients can enforce constraints on the minimum required server version
- Servers can enforce constraints on the minimum required client version

Doing this is a bit challenging. Both sides need to know what the other
side's version is - thankfully, we've been doing this for a *while*.

- The client can get the server version through the Buildkit client.Info
  call (where we've been attaching this data since v0.9.0).
- The server can get the client version info through old telemetry
  labels (after explicitly handling old-style telemetry labels), or
  through a newly introduced ClientMetadata field.

With this new data, we can enforce version constraints! However, in the
case of the server-side version constraints for the client, we hit a
slightly weird case - the client got stuck in a retry loop when the
Session failed to start. This isn't quite right behavior anyways (any
error from the main Session call that starts a server wouldn't cause an
immediate failure), it just appeared when we try and display version
constraints. We can easily fix this going forwards - on older clients,
we just get the failure logs, the retry will still occur, and we fail
about 5 minutes.

Signed-off-by: Justin Chadwell <me@jedevc.com>

* tests: add test for checking version compatibility

Signed-off-by: Justin Chadwell <me@jedevc.com>

* docs: add backwards compatability notes to RELEASING

Signed-off-by: Justin Chadwell <me@jedevc.com>

---------

Signed-off-by: Justin Chadwell <me@jedevc.com>