engine: add logging to help debug "transport not supported"
This has been reported to happen intermittently by a few users. The
error originates from gqlgen and suggests that something in the http
request is unhandled; it basically tries to find a matching pattern for
gql requests based on http method and some headers.

It's unclear how this could happen only occasionally and I've never seen
this myself yet, so I think our best bet is to log the parts of the
request checked by gqlgen and wait for it to get hit again.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>