ci: enable best-effort cache reuse for nested engines
We were previously always putting a random id in the cache volume used
for engines created in various CI functions, which meant none of them
ever had any local cache. The main reason was that two engines can't use
the same local cache at the same time.

We can't lift that limitation, but by using the engine version and
private cache mounts we can get *best effort* re-use of these cache
volumes, which is a simple enough performance improvement for some cases
that we may as well take it.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>