client: set a minimum connect timeout for buildkit
we're changing the default minConnectTimeout and backoff settings since
in some cases, the `connect` happens before the engine is ready to
accept connections, causing the client to wait for the default
minConnectTimeout (20s) before retrying. This causes very long
connection times when the engine is not ready to accept connections.

fixes #7599

Signed-off-by: Marcos Lilljedahl <marcosnils@gmail.com>