engine: disentangle client+session state from Query
To fully isolate client-specific resources within a session, we need to
move off of storing things like the SecretStore as fields on Query. This
is because the dagql cache is shared across a session, meaning it can
return objects from other clients that should not be used.

I just went ahead and got rid of all the fields on Query and turned them
into methods. *Technically* it wasn't necessary for all of them, but
it's more consistent this way and sets us up for cross-session dagql
caching if ever needed.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>