chore: isolate modulesourcekind to its own subpackage of core
The CLI now relies on modulesourcekind to isolate the kind of module
being loaded. This helps us have a consistent behavior regarding
local file / dir refs that resembles remote ones

However, as the enum implementation was part of the core package, it
introduced build constraints. Moving the implementation to the core/modules subpackage solves it

Signed-off-by: Guillaume de Rouville <guillaume@dagger.io>