ci: fix git publish tag confusion (#7648)
This is fairly fiddly to do correctly. Essentially, we need to be able
to handle at least these inputs:
- Branch names, like `main`
- Tag names, like `v0.11.7`
- Full refs, like `refs/pull/<pr>/merge`

We can't simply use `git fetch <ref>`, since that doesn't create a local
ref copy. We shouldn't use a refspec that mirrors to a single name,
since that strips `refs/tag/...` which means tags are lost.

The correct way to do this is using a `git fetch` refspec, to mirror all
remote refs that match the target name to a local ref. Then this
satisfies all the above inputs correctly.

Additionally, for dry-runs, I've added a `rev-parse` sanity check, so
can check the results more easily.

Signed-off-by: Justin Chadwell <me@jedevc.com>