fix: prevent infinite progress loop with forced simplification
Forced simplification could occasionally produce a weird infinite loop:

- `moduleSource.resolveDependency` is an identity function (for git
  sources), taking an argument and returning it unmodified.
- Arguments in a call to `moduleSource.resolveDependency(dep: ...)`
  (from an unrelated function) will be forced to be simplified, since
  it's an internal function.
- There are no reasonable simplifications for the `dep` argument here -
  so since it must be simplified, it is simplified to
  `moduleSource.resolveDependency(dep: ...)` (since `resolveDependency`
  is an identity function.
- This recurses infinitely, with `dep` being expanded to it's parent.

The solution here is to never allow simplification to a parent like
this. We can detect this case by inspecting the arguments to a candidate
simplification, and ensuring that it doesn't ever accept the original
call - if it does, then this clearly isn't a "real" simplification, and
is actually a complexification by selecting a parent.

Signed-off-by: Justin Chadwell <me@jedevc.com>