fix: be less aggressive with forced simplification
Forced simplification is tricky - yes, the force argument was able to
return invalid simplifications, but also, we've been setting it too
aggressively.

We should *only* be applying forced simplification at the boundary of
where the public and internal APIs meet - once we're inside the internal
API, we shouldn't force simplification.

This is an alternate fix to the issue, but it makes sense to fix in both
places.

Signed-off-by: Justin Chadwell <me@jedevc.com>