Use dynamically sized dind runners for SDK and testdev (#7644)
* Use dynamically sized dind runners for SDK and testdev

As was discussed here:
https://github.com/dagger/dagger/pull/7483#discussion_r1636854755, since
[#7628](https://github.com/dagger/dagger/pull/7628) got merged we are
using dind runners for SDK jobs. Default dind runners have 32c and 16Gi,
which for SDK jobs is excessive. Since we now have support for
dynamically sized dind runners, we migrate SDK jobs to 4c (like we used
to run it) and we leave testdev with 32c jobs.

Signed-off-by: Matias Pan <matias@dagger.io>

* ci: allow custom dev container+output names

Signed-off-by: Justin Chadwell <me@jedevc.com>

---------

Signed-off-by: Matias Pan <matias@dagger.io>
Signed-off-by: Justin Chadwell <me@jedevc.com>
Co-authored-by: Justin Chadwell <me@jedevc.com>