Publish Wolfi Engine variants - with & without NVIDIA drivers support
This publishes two new image variants:
1. *-wolfi
2. *-wolfi-gpu

We want to take this through the paces - automated & manual tests - and
check if it's feasible to consolidate all images into a single one:
Wolfi with NVIDIA.

I have also made it clearer in workflow descriptions that this isn't a
GPU image variant. While NVIDIA is the dominant GPU vendor today, Intel
& AMD are still relevant and in some cases ahead (e.g. Aurora & Frontier
supercomputers). A bunch of homelabbers (including myself) have Intel
GPUs which we would expect the GPU variant to just work with. I don't
think that's the case today.

While this doesn't close https://github.com/dagger/dagger/issues/5668,
it's the second step in that direction.

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>