fix(ts-sdk): correctly resolve non-primitive default value
There was an issue that created an error when a non-primitive default
value was set to a function like

```typescript
@object()
class Default {
  @func()
  a(
    dir: Directory = dag.directory()
      .withNewFile("/foo", "hello world")
    ): Container {
    return dag.container().withMountedDirectory("/dir", dir)
  }
}
```

This commit fixes the issue to ignore the default value registration
if it's not a primitive type and let the TypeScript runtime resolve
it during the execution.

It's a workaround to accomplish the result because resolving the
value during the registration requires to run the TS code in a custom
runtime that must contains all user's dependency which may be
resource consumming + complex to resolve.

Signed-off-by: Tom Chauveau <tom@epitech.eu>