fix: ensure gc policy percentage is correct
When originally updating this, we only updated the default CLI option -
however, this is confusing, and isn't exactly how that works.

Because we use `GlobalIsSet`, this default is never actually the one
that's used. So, we can refactor this, and set a global const to define
the default gc policy ourselves, outside of buildkit.

This also opens us up to modifying the gc policy in more detail in the
future.

Signed-off-by: Justin Chadwell <me@jedevc.com>