avoid massive OTel test log volume
Rather than logging to the test span, log to the underlying test logger.

When we log to the test span we're unconditionally sending what is in
total a massive amount of log volume. Previously we only saw these logs
on test failure, so this change brings us back to that by sending them
to the underlying *testing.T logger instead.

Signed-off-by: Alex Suraci <alex@dagger.io>