add `testctx` for instrumenting integration tests with OTel (#7532)
* core/integration: use testctx for ctx propagation

* testing.T -> testctx.T
  * parallel by default (currently no way _not_ to)
  * carries around a context.Context
  * supports middleware patterns, e.g. for OTel instrumentation
  * embeds *testing.T, so fully compatible; if you need the concrete
    type, just access the field
* tests now live on a Suite

Signed-off-by: Alex Suraci <alex@dagger.io>

* fix WithTimeout being applied too early

ended up timing out while waiting on t.Parallel

Signed-off-by: Alex Suraci <alex@dagger.io>

* fix TestDirectoryWithTimestamps flakiness

apparently the second column is number of hardlinks; we dont care

Signed-off-by: Alex Suraci <alex@dagger.io>

* fix TestCacheVolume parallelism

it expected t.Run to run in series

Signed-off-by: Alex Suraci <alex@dagger.io>

* appease linter

Signed-off-by: Alex Suraci <alex@dagger.io>

* fix parallel-unfriendly workdir/export tests

Signed-off-by: Alex Suraci <alex@dagger.io>

* fix ContainerSuite.TestExport parallelism

Signed-off-by: Alex Suraci <alex@dagger.io>

* fix ctx propagation

Signed-off-by: Alex Suraci <alex@dagger.io>

* more generous timeout for MultiSameTrace

Signed-off-by: Alex Suraci <alex@dagger.io>

* remove gotestsum

something is hiding race detector output. the new instrumentation
replaces the need for prettified output and slowness reports, so let's
just eliminate a variable.

Signed-off-by: Alex Suraci <alex@dagger.io>

* bump ExitsZeroOnSignal timeout

failed twice, but looks like our infra is just slow

Signed-off-by: Alex Suraci <alex@dagger.io>

* DNM: github: dagger call with -s

just to see if it makes any difference

Signed-off-by: Alex Suraci <alex@dagger.io>

* fix accidental var reuse in LICENSE tests

Signed-off-by: Alex Suraci <alex@dagger.io>

* Revert "DNM: github: dagger call with -s"

This reverts commit 8951032ecec8bf194eeb7eb1b106fcd57926b57b.

Signed-off-by: Alex Suraci <alex@dagger.io>

* avoid massive OTel test log volume

Rather than logging to the test span, log to the underlying test logger.

When we log to the test span we're unconditionally sending what is in
total a massive amount of log volume. Previously we only saw these logs
on test failure, so this change brings us back to that by sending them
to the underlying *testing.T logger instead.

Signed-off-by: Alex Suraci <alex@dagger.io>

---------

Signed-off-by: Alex Suraci <alex@dagger.io>