ci: Tag Engine image with main when publishing from the main branch
We want this so that we can reference `registry.dagger.io/engine:main`
in our CI, and have our runners pick up the latest Engine Image (a.k.a.
dev). I didn't want to call this tag `dev` since:
- it assumes there is a single dev branch (unlikely to be true for long
  based on the latest releasing discussions)
- it requires us to make an extra mental jump from dev to main to
  understand it's true origin
- it required me to change existing code (we are already using the dev
  concept when building from local)

I didn't want to spend too much time on this since more changes related
to the base image are comming right after this (the PR is mostly
finished & it builds on top of this one). These will affect all Engine
images that we currently publish: Alpine (the default one), Wolfi & GPU.
There is a refactoring opportunity with VersionInfo, but I wanted to
keep it low risk & not change any of the existing behaviour - this is an
additive change, on top of what we were already doing.

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>